+++ packages/leancode_contracts_generator/lib/src/generator_script.dart
@@ -35,7 +35,7 @@ class GeneratorScript {
   final List<String> args;
   final String? overrideGeneratorVersion;
 
-  static const supportedVersion = '0.1.0-alpha11';
+  static const supportedVersion = '1.0.0';
 
   static const String _script = _$script;